% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTx.R
\name{LT_probit}
\alias{LT_probit}
\title{Lethal Time Probit}
\usage{
LT_probit(
  formula,
  data,
  p = NULL,
  weights = NULL,
  subset = NULL,
  log_base = NULL,
  log_x = TRUE,
  het_sig = NULL,
  conf_level = NULL,
  conf_type = NULL,
  long_output = TRUE
)
}
\arguments{
\item{formula}{an object of class \code{formula} or one that can be coerced to that class: a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which \code{LT_probit} is called.}

\item{p}{Lethal time (LT) values for given p, example will return a LT50 value if p equals 50. If more than one LT value wanted specify by creating a vector. LT values can be calculated down to the 1e-16 of a percentage (e.g. LT99.99). However, the tibble produced can and will round to nearest whole number.}

\item{weights}{vector of 'prior weights' to be used in the fitting process. Only needs to be supplied if you are taking the response / total for your response variable within the formula call of \code{LC_probit}. Otherwise if you use cbind(response, non-response) method you do not need to supply weights. If you do the model will be incorrect. If you don't supply weights there is a warning that will help you to make sure you are using one method or the other. See the following StackExchange post about differences \href{https://stats.stackexchange.com/questions/259502/in-using-the-cbind-function-in-r-for-a-logistic-regression-on-a-2-times-2-t}{cbind() function in R for a logistic regression}.}

\item{subset}{allows for the data to be subseted if desired. Default set to \code{NULL}.}

\item{log_base}{default is \code{10} and will be used to  calculate results using the anti of \code{log10()} given that the x variable has been \code{log10} transformed. If \code{FALSE} results will not be back transformed.}

\item{log_x}{default is \code{TRUE} and will calculate results using the antilog of determined by \code{log_base} given that the x variable has been \code{log()} transformed. If \code{FALSE} results will not be back transformed.}

\item{het_sig}{significance level from person's chi square goodness-of-fit test that is used to decide if a heterogeneity factor is used. \code{NULL} is set to 0.15.}

\item{conf_level}{Adjust confidence level as necessary or \code{NULL} set at 0.95.}

\item{conf_type}{default is \code{"fl"} which will calculate fudicial confidence limits per Finney 1971. If set to \code{"dm"} the delta method will be used instead.}

\item{long_output}{default is \code{TRUE} which will return a tibble with all 17 variables. If \code{FALSE} the tibble returned will consist of the p level, n, the predicted LC for given p level, lower and upper confidence limits.}
}
\value{
Returns a tibble with predicted LT for given p level, lower CL (LCL), upper CL (UCL), LCL, Pearson's chi square goodness-of-fit test (pgof), slope, intercept, slope and intercept p values and standard error, and LT variance.
}
\description{
Calculates lethal time (LT) and
its fiducial confidence limits (CL) using a probit analysis
according to Finney 1971, Wheeler et al. 2006, and Robertson et al. 2007.
}
\examples{
head(lamprey_time)

results <- LT_probit((response / total) ~ log10(hour),
p = c(50, 99),
weights = total,
data = lamprey_time,
subset = c(month == "May"))

# view calculated LT50 and LT99 for seasonal
# toxicity of a piscicide, 3-trifluoromethyl-4-nitrophenol, to lamprey in 2011

results

# dose-response curve can be plotted using 'ggplot2'
}
\references{
Finney, D.J., 1971. Probit Analysis, Cambridge University Press, Cambridge, England, ISBN: 052108041X

Wheeler, M.W., Park, R.M., and Bailey, A.J., 2006. Comparing median lethal concentration values using confidence interval overlap or ratio tests, Environ. Toxic. Chem. 25(5), 1441-1444.10.1897/05-320R.1

Robertson, J.L., Savin, N.E., Russell, R.M. and Preisler, H.K., 2007. Bioassays with arthropods. CRC press. ISBN: 9780849323317
}
